Refact: Move image-related API calls to a separate service file for better reusability and cleaner code structure in AddPattern.tsx & PatternDetailPage.tsx
